fix(botcore): Systematic analyzer fixes for Integration, Patterns, Features, Market, Configuration, Extensions, HealthChecks, Fusion, and StrategyDsl folders (#448)
* Initial plan

* fix(strategyDsl): Fixed CS1519 syntax error in ExpressionEvaluator.cs - removed extra closing brace


* fix(integration): Fixed CA1819 and CA1002 - Changed array/List properties to IReadOnlyList in YamlSchemaValidator and ShadowModeManager


* fix(integration,features,healthchecks): Fixed S3267 - Simplified loops using LINQ Where/All/Count methods


* fix(market,healthchecks,patterns,fusion): Fixed CA1869 and S1905 - Cached JsonSerializerOptions and removed redundant casts


* fix(strategydsl,market,integration): Fixed CA2016, S1121, S1066 - Forward cancellation tokens and simplified nested conditions


* fix(features,healthchecks): Fixed CA2254 and CA2234 - Use structured logging and Uri overloads


* chore(agent5): Update status - 46 violations fixed across 6 batches (23% of target)


* docs(agent5): Update Change-Ledger with Round 206 - 46 violations fixed across 6 batches


---------

+++ src/BotCore/Patterns/Detectors/CandlestickPatternDetector.cs
@@ -351,7 +351,7 @@ public class CandlestickPatternDetector : IPatternDetector
         if (range == 0) return 0.0;
         
         var bodyRatio = bodySize / range;
-        return bodyRatio > MinSpinningTopBodyRatio && bodyRatio < MaxSpinningTopBodyRatio ? (double)SpinningTopConfidence : 0.0;
+        return bodyRatio > MinSpinningTopBodyRatio && bodyRatio < MaxSpinningTopBodyRatio ? SpinningTopConfidence : 0.0;
     }
 
     private static double DetectThreeConsecutive(IReadOnlyList<Bar> bars, bool bullish)
